+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * Improve text in README and bayesSSM vignette.
 
+* Fixed a bug in stratified resampling.
+
 # bayesSSM 0.4.7
 
 * Initial CRAN submission.